+++ bin/varnishtest/vmods.h
@@ -30,3 +30,4 @@
 VTC_VMOD(std)
 VTC_VMOD(debug)
 VTC_VMOD(directors)
+VTC_VMOD(vtc)